build: prefer gcc over tcc
as the conditional statements seem to be broken for now in tcc for some reason, i'll use gcc for now. the line that caused the switch was this one: `cam_speed = (IsKeyDown(KEY_LEFT_SHIFT)) ? CAMERA_MAX_SPEED : CAMERA_MOVE_SPEED;` , it was always evaluating `true` and was setting the `cam_speed` to `CAMERA_MAX_SPEED`; no other way of setting `cam_speed` seems to be working either on tcc.