feat: add a way to save the current map
press CTRL+S to save the current map with all of its objects and their locations. Use the built example `mapreader` (examples/readmap.c) to read the map data. Currently no way to select the name for the mapfile, so it's always "test_map.lol" in the same location the map editor was opened in