build: allow for a non-global installation of dependencies
this is a much better solution for those who don't want to accidentally BREAK something on their system without even realizing it. instead of doing something system-wide it's much better to localize the possible damages to only the project which depends on certain things. yes, now you have to do ./dep_autoinstall for each new computer on which you compile the editor, but this DOES bring a lot of safety guards in order for people to not somehow brick their systems. safety first when it comes to libraries on the system