feat: add new field "data" for objects
this allows for integration with virtually any existing game engine, but it also simplifies creation of scenes for everyone. E.G. : to create a light source, specify label as "LIGHT", then, in your game engine when reading the data simply look at the label section, if it's "LIGHT", then you can render a light source at the location of the object! :) the example is extremely simplified, but it's only one of infinite possibilities you can have with this field! TODO: fix a bug for when the field is edited via gui but the user doesn't press enter yet and then the user clicks on another object, resulting in a weird state for both of them