perf: remove redundant shader value being set
since screen_w and screen_h are global variables and the shader values are set via their addresses, we don't have to re-set the shader values as those already point to our variables. we can just change the global variables and those will change within the shader as those are uniforms!