fix: get rid of segfaults for object selection
this is done by completely revising the selection system, with num_selected_objects going up to (ObjectCounter)-1 when none are selected, up to 0 when one is selected and so on. this also fixes a weird issue where selecting objects with mouse doesn't segfault but with gui does? not sure how that was working, in the end i just made a function and the issue seemed to have gone away. fixed issue where memory reserve for selected objects was 1 off, making so the first rellocation IS successful, but the second one was unreachable because num_selected_objects was 19 (not divisible by 10, i.e. SELECTED_OBJECTS_MEMORY_RESERVE)